requirements.txt
bleach==3.1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.0.6
django-allauth==0.41.0
django-appconf==1.0.2
django-ckeditor==5.9.0
django-crispy-forms==1.7.2
django-etc==0.10.0
django-formtools==2.1
django-geoposition==0.3.0
django-haystack==2.6.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-registration-redux==2.4
django-rest-auth==0.9.5
django-select2==6.1.0
django-sparklines==0.1
django-todo==2.4.9
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
fanstatic==1.1
idna==2.9
jieba==0.38
js.jquery==3.3.1
Markdown==2.6.8
oauthlib==3.1.0
Pillow==5.2.0
psycopg2==2.8.5
pycparser==2.20
Pygments==2.2.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.5
requests==2.23.0
requests-oauthlib==1.3.0
shutilwhich==1.1.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
Unidecode==1.1.1
urllib3==1.25.8
webencodings==0.5.1
WebOb==1.8.6
whitenoise==5.0.1
Whoosh==2.7.4
